+++ pbusinessaccount.h
@@ -39,8 +39,8 @@ int do_account_modifyshare(psync_shareid_t usrshrids[], uint32_t uperms[], int n
 int do_account_users(psync_userid_t iserids[], int nids, result_visitor vis, void *param);
 int do_account_teams(psync_userid_t teamids[], int nids, result_visitor vis, void *param);
 
-void get_ba_member_email(uint64_t userid, char** email /*OUT*/, int *length /*OUT*/);
-void get_ba_team_name(uint64_t teamid, char** name /*OUT*/, int *length /*OUT*/);
+void get_ba_member_email(uint64_t userid, char** email /*OUT*/, size_t *length /*OUT*/);
+void get_ba_team_name(uint64_t teamid, char** name /*OUT*/, size_t *length /*OUT*/);
